chore: move section components into sections folder
