+++ src/App/App.js
@@ -2,16 +2,18 @@ import React from 'react'
 import { HashRouter as Router, Route, Switch, Redirect } from 'react-router-dom'
 import { CssVariables } from '@dhis2/ui'
 
+import {
+    Dashboard,
+    DataExportForExcel,
+    OutliersAndMissingDataAnalysis,
+    ConsistencyAnalysis,
+    AnnualReport,
+    Administration,
+} from '../sections'
+
+import * as PATHS from '../config/paths.js'
 import SideNav from '../components/SideNav.js'
 import NoMatch from '../components/NoMatch.js'
-import Dashboard from '../Dashboard/Dashboard.js'
-import OutliersAndMissingDataAnalysis from '../OutliersAndMissingDataAnalysis/OutliersAndMissingDataAnalysis.js'
-import ConsistencyAnalysis from '../ConsistencyAnalysis/ConsistencyAnalysis.js'
-import AnnualReport from '../AnnualReport/AnnualReport.js'
-import Administration from '../Administration/Administration.js'
-import DataExportForExcel from '../DataExportForExcel/DataExportForExcel.js'
-import * as PATHS from '../config/paths.js'
-
 import classes from './App.module.css'
 
 const App = () => (
