+++ src/sections/index.js
@@ -1,0 +1,15 @@
+import Administration from './Administration/Administration.js'
+import AnnualReport from './AnnualReport/AnnualReport.js'
+import ConsistencyAnalysis from './ConsistencyAnalysis/ConsistencyAnalysis.js'
+import Dashboard from './Dashboard/Dashboard.js'
+import DataExportForExcel from './DataExportForExcel/DataExportForExcel.js'
+import OutliersAndMissingDataAnalysis from './OutliersAndMissingDataAnalysis/OutliersAndMissingDataAnalysis.js'
+
+export {
+    Administration,
+    AnnualReport,
+    ConsistencyAnalysis,
+    Dashboard,
+    DataExportForExcel,
+    OutliersAndMissingDataAnalysis,
+}
